requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.3
python-dateutil==2.8.2
setuptools==67.6.1
Pillow==10.2.0
